chore: move GrpcContext from grpc package to auth
The layered architecture tests prohibit the access of the grpc layer from the service layer. Therefore, the GrpcContext needs to be moved out of the grpc layer into an unrestricted layer.
